checkLvl, checkIV, calculateDamage, returnEffAtk splitted
