+++ interface/src/js/Damage/calculateDamage.js
@@ -1,0 +1,6 @@
+export function calculateDamage(movePower, aAttack, dDefence, multiplier) {
+    if (aAttack === 0 || dDefence === 0 || multiplier === 0) {
+        return 0
+    }
+    return Math.trunc(movePower * 0.5 * (aAttack / dDefence) * multiplier + 1)
+}
